Use unsigned integer types for non-negative counts (nsub, nsim, nall, nSize)

+++ inst/include/rxode2parseStruct.h
@@ -240,17 +240,17 @@ typedef struct {
 typedef struct {
   rx_solving_options_ind *subjects;
   rx_solving_options *op;
-  int nsub;
-  int nsim;
+  uint32_t nsub;
+  uint32_t nsim;
   int neta;
   int neps;
   int nIndSim;
   int simflg;
-  int nall;
+  uint32_t nall;
   int nevid9;
   int nobs; // isObs() observations
   int nobs2; // evid=0 observations
-  int nr;
+  int64_t nr; // int64_t because nobs*nsim can exceed INT_MAX in VPC with many subjects/timepoints/sims
   int add_cov;
   int matrix;
   int needSort;
